requirements.txt
uvloop==0.5.4
msgpack-python==0.4.8
nose==1.3.7
